Implements rating specs
